Optimize enum usage
* Reduce enum storage size
* Convert HasFlag() calls to direct bitwise comparison operator

+++ src/AngleSharp/Css/Values/Number.cs
@@ -133,7 +133,7 @@
         /// <summary>
         /// An enumeration of angle representations.
         /// </summary>
-        public enum Unit : ushort
+        public enum Unit : byte
         {
             /// <summary>
             /// The value has been given as an integer.
