+++ src/AngleSharp/Dom/Constants/FilterResult.cs
@@ -6,7 +6,7 @@
     /// An enumeration that contains the valid results of examining a node.
     /// </summary>
     [DomName("NodeFilter")]
-    public enum FilterResult : ushort
+    public enum FilterResult : byte
     {
         /// <summary>
         /// The Node is accepted.
